Update patch set 7
Patch Set 7:

So I tried using fixtures to make this easier to read and I think it came out much worse. merger.Repo is wrapping a git.Repo. It ensures that it is cloned and for operations run on that cloned repo it uses a context manager to keep track of when to GC the internal repo refernce. When you pull that context manager out into a fixture you suddenly have an additional class with mutliple methods that need to call into the ensure cloned logic of merger.Repo which means you are calling into the fixture then back out again. I think having the context manager in merger.Repo makes sense as a result. And finally if using the fixture as a context manager when using the fixture there isn't any difference with the existing code that uses a contextlib decorated function.

It has been a long day though and I may have completely looked at this the wrong way. Are there specifics on how you think this will make the code easier to read?

Patch-set: 7
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0
Label: Workflow=0