Raise an exception on an invalid reporter
An actual config file error should be caught by the validator, so
this should never happen, however, a code error in Zuul was causing
this error to be logged on layout testing (even though the config
file was correct).  See https://review.openstack.org/55184

This will raise an exception in that situation and hopefully prevent
such a coding error in the future.  Also, since it's an error
condition that shouldn't ever happen (thanks to layout validation),
it truly merits an exception.

Change-Id: I681f45656504b2bd00cface5c2a4607eab0155ae