Update patch set 11
Patch Set 11: I would prefer that you didn't merge this

OK, so I spotted some errors in the tests and had to sort of fix them interactively, so I went ahead and pushed a new patchset with working tests.

However, I was already leaning toward the idea that this should be a required parameter as well, and now I see other people in the comments feel that way.

So why don't we go ahead and make it required?  That will simplify things somewhat -- especially the negative test which is somewhat problematic as you'll note.

Patch-set: 11
Label: Code-Review=-1