Update patch set 13
Patch Set 13: I would prefer that you didn't merge this

(5 inline comments)

The option should be documented in zuul.rst as well.  Also, left a few notes of things that can be removed if we're going to require the param.  And finally, if you'd like to propose a commit to infra/config to add zuul_url that would be nice.

Patch-set: 13
Label: Code-Review=-1
Label: Workflow=0