Update patch set 18
Patch Set 18: I would prefer that you didn't merge this

I would prefer that the new option be optional. This will make it easier for us to deploy the new zuul code and be friendly to other users of zuul because we will maintain backward compat with older versions of zuul configs.

At the very least we will need an openstack-infra/config change that adds the new option, if it is not made optional.

Patch-set: 18