Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

(7 inline comments)

I have a few things I think should be fixed before landing inline.

On top of that, I think this deserves tests : you have changes to the core of zuul which is well tested, we should maintain that coverage.

Lastly, we have a pattern in OpenStack of having separate client libraries for this sort of thing - a python-zuulclient would make a lot of sense to me here.

Patch-set: 2
Label: Code-Review=-1