Update patch set 2
Patch Set 2: (6 inline comments)

Agreed on tests.

Unless you have a _very_ permissive production environment (we don't), the client is unable to be used except on the server host due to the complete lack of authn/z.  This is basically an admin tool to be run on-server, so I think it makes sense to keep it here for now.

Patch-set: 2
Label: Code-Review=0