make enqueue_time durable by caching in change
Once a change has an enqueue time, we should keep it, regardless
of how it moves around in the queue. So store the enqueue_time back
in the change itself when we create it for the item. If the change
has an enqueue_time, then use that instead of generating a new one.

This should address the issue where enqueue_times get lost when
we do a promote of the queue.

Change-Id: I4dce23d00128280c7add922ca9ef5018b57d1cf3