Die out on invalid layout
I spend a while figuring why Zuul was not properly initializing and
never quitting.  The root cause was that my layout.yaml was invalid and
Zuul (in foreground mode with -d) did not exit.

This patch will cause Zuul to exit whenever the sceduler can not be
properly configured.  We might want to do that before starting the
internal Gearman server.

Change-Id: Ic4b9803611049d30bbc29b678c7bb8b447b679ef