Update patch set 3
Patch Set 3:

I've been starting to think that it might be desirable to allow some caching.  Is it possible to set headers to keep a cache for 5 or 30 seconds for instance?  Is that desirable, and if so, should we make this a config option?

(I'm imagining a situation where there are lots of viewers of the zuul status page and zuul essentially wastes effort rebuilding the same json many times every few seconds.)

Patch-set: 3
Label: Code-Review=0
Label: Workflow=0