Update patch set 3
Patch Set 3: I would prefer that you didn't merge this

I think we should support both rather than removing support for HTTP. There are a couple of reasons for this. It removes a feature that people may rely on despite needing an ssh connection anyways. And the reason they may rely on it is there are only so many threads available to service ssh connections.

Finally, it would be nice to get this into gerritlib and make zuul use gerritlib, but that doesn't need to be done to support this change.

Patch-set: 3
Label: Code-Review=-1
Label: Workflow=0