Update patch set 3
Patch Set 3:

I think James brings up a good point. I'm assuming infra would then put this behind varnish or something?

Keep in mind that the value of the cache will have a nyquist effect for the viewer. ie, a 5sec cache with a 5sec refresh could result in a 10sec old page. Clearly not an issue, but if we are hard coding the values it's something to keep in mind.

Patch-set: 3
Label: Code-Review=0