Update patch set 3
Patch Set 3:

At Wikimedia, I have migrated the CI box behind a shared varnish box to reclaim a public IP address. Caching is a side effect.

On our setup we mostly have short jobs running and people load the status page to look at the progress while waiting.  So adding a delay there via caching would be disruptive.

With Cache-Control: no-cache, the client revalidates with varnish and varnish revalidates with the server. That does not save any load on Zuul webapp side when there is a swarm of people.

What could be done is to instruct varnish to serve a copy but still get client to revalidate.  That is done by using "max-age: 0" which is for client and shared cache, then "s-maxage: 5" which would override the max-age value for shared cache.  Hence the resulting header would be:

 Cache-Control: maxage=0; s-maxage=5

BUT, if the Zuul web app is not behind a shared proxy, all clients will still hit it on each refresh (because of maxage=0) :-/

Patch-set: 3