Update patch set 10
Patch Set 10: I would prefer that you didn't merge this

(3 inline comments)

Some inline comments.

Also, I think that this should be opportunistically tested if the "openstack_citest" database is available (it is on the jenkins slaves).  I think that python-mysqldb should be in test-requirements but not in requirements (it should be optional).  We should probably have an import test to determine whether the reporter should be available.  And finally, it should be documented.

I'd also recommend getting at least one mysql expert to review this as well (such as devananda or mordred).

Patch-set: 10
Label: Code-Review=-1
Label: Workflow=0