Update patch set 3
Patch Set 3: I would prefer that you didn't merge this

I agree with jhesketh's comments. This change and the one after it should be squashed together so that this change is appropriately tested.

Also, this change will conflict with https://review.openstack.org/#/c/71627/3/zuul/scheduler.py. Might be worth coordinating with jeblair so that stack can be shifted as appropriate.

Patch-set: 3
Label: Code-Review=-1
Label: Workflow=0