Update patch set 7
Patch Set 7: I would prefer that you didn't merge this

(4 inline comments)

First, this is great; I think we're 95% of the way there in one patch.  I left a few things inline.

One use case that we may have is to support multiple containers.  For instance, all jobs will need to upload logs, but some jobs may also need to upload artifacts such as tarballs.  For those, we may want to provide two hmac signatures for two different containers.  And then why stop there?

We could take an approach similar to what we're doing now and handle that in configuration-code (the custom functions that we use to set per-job environment variables) so that you can write a custom function that calls the swift variable setup again for a tarball job (with different variable names).

Or we could make that first-class yaml configuration.  So perhaps for a job (or meta-job) you would say something vaguely like:

jobs:
  - name: '^.*$'
    swift:
      - container: logs
        refspec-path: '{change.number}/...'
        ref-path: '{change.newrev}/...'
        other-path: '{pipeline.name}/...'
  - name: '^.*-tarball$'
    swift:
      - container: tarball
        pre-merge-path: '{job.name}/'

So every job gets the log container info (with a different path depending on whether the change is pre, post, or periodic), and tarball jobs additionally are allowed to upload to the tarball container.

Actually I don't like the fact that the log paths are defined three times for global meta-job; perhaps we should add something to the model where a changeish tries to provide a descriptive path itself, then we can just say "path: '{change.path}/...'".

Patch-set: 7
Label: Code-Review=-1
Label: Workflow=0