Update patch set 6
Patch Set 6: I would prefer that you didn't merge this

(1 inline comment)

-1 for the prettytable version thing

This is just for discussion:

This change looks great, except one thing occurs to me -- there are now two places in the code where we're building JSON representation of the current running state.  Why not add the missing info to the status json and then re-use that for the command line version?

Patch-set: 6
Label: Code-Review=-1
Label: Workflow=0