Update patch set 6
Patch Set 6: (1 inline comment)

To me it makes more sense to retrieve the status information via the RPC now that it exists. I would suggest that the status.json feed should retrieve its information over the RPC as well but I fear gearman may be too slow. However we could move towards not having a web server as part of zuul (although perhaps keep it in contrib/) and have a script that consumes, caches and serves the RPC results appropriately.

Patch-set: 6