Update patch set 7
Patch Set 7:

I think I may have been unclear in the last comment.  I wasn't suggesting that you not provide the info via the RPC layer, but instead I was just suggesting that you not build another JSON blob representing more or less the same info.

In other words, add the missing info to the JSON builder used by the web service, then call that from the RPC layer to get the JSON to pass over the wire.

Basically, I'm thinking that when the JSON changes, it would be nice to avoid having to change it in a couple of places.

Patch-set: 7
Label: Code-Review=0
Label: Workflow=0