Update patch set 3
Patch Set 3: Looks good to me (core reviewer)

Not approving as I want to review the whole stack before I start hitting the go button.

Also, this is a potentially scary change to land, will definitely want to keep an eye on zuul when this goes in.

Patch-set: 3
Label: Code-Review=+2
Label: Workflow=0