Update patch set 11
Patch Set 11:

Okay, so I can't get this race to reproduce locally.

Basically the problem is that the gearman server hasn't told the gearman client about the updated work data before the check (sometimes). I can't see an easy way to reach into the client to check if the packet has been returned so instead we'll wait for the build information to be updated.

Patch-set: 11
Label: Code-Review=0