Add a remote url override location
Zuul constructs a git clone and fetch url from its gerrit connection
information. There may be cases where zuuld can connect to gerrit for
events and git info but the zuul merge workers cannot. Add a config
option to override the git location url for zuul merge workers so that
they can talk to gerrit replicas instead of gerrit directly. Because the
merge workers are not talking to gerrit directly also add an optional
poll for ref existence in the merge worker.

To test this make it possible to subclass TestScheduler and feed CONFIG
values into the ConfigParser prior to running TestScheduler's setUp.
This means all scheduler tests can be run under subclasses with
different configs.

Change-Id: I3e7f2ab07051610e8554555a7c5fecd458e38330