Add queue-name parameter to projects
Allow an optional queue-name parameter to be set for a project.
As that project is combined with others into shared change queues,
give the queue that name.  This allows us to, say, set the queue
name of tempest to 'integrated' and end up with the shared change
queue of all the OpenStack integrated projects named 'integrated'.

With that, we can do things like emit stats for the 'integrated'
queue.

Change-Id: I2c7ac81e7642dedd7e1147c1ec4382fc414c5420