Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

(6 inline comments)

Looks good. Some nit picking and bike-shedding within.

Also, as mentioned on IRC, perhaps the queue name should be assigned to a job not a project since it is the job that is used to tie together a queue.

Patch-set: 2