Add configurable footer-message reports
Currently there is no way to add extra information to the end of a
report sent back by zuul. This adds a 'footer-message' field to the
pipelines allowing operators to leave extra information at the end
of a build result.

This is useful for leaving CI contact and debugging information.

Change-Id: Ieae62e845915fa3997353b6b425e215966a9338c