Update patch set 1
Patch Set 1:

Joshua wrote:
> zuul is meant to be independent from OpenStack. A generic gating program per se

Yeah that is what I thought. I haven't found a way to feature switch oslosphinx so it only get applied / configured when running on the OpenStack Jenkins which generates the documentation at ci.openstack.org.

Maybe it can be done with tox env and a command to alter the list of Sphinx extension + add the oslosphinx as a dependency.

Patch-set: 1