Update patch set 4
Patch Set 4:

Clark wrote:
> Why would you put zuul behind a cache? I can't think of a reason to do that, but there seems to be one which has made me curious.

I explained it as a comment on patchset 3 above. Basically it is to reclaim a public IP, Zuul is on an internal IP and its status.json is exposed via a Varnish shared by multiple services.

I got to find sometime this week to complete this feature. Will add yet another setting to finely adjust / disable the caching TTL.

Patch-set: 4