Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

(4 inline comments)

This is mainly a straw man, intended to point out where the hard-coded timeout currently resides. If we keep this check, it should ideally be turned into a configurable value and updated on any reconfig.

It was also suggested that we can just remove the replication check entirely, but the more I think about it I wonder whether that leaves us more likely to hit a race condition where a job starts and updates from a (non-Zuul, branch tip) reference on a remote Git mirror and doesn't actually get the change which just merged. For example, consider that we're testing changes for projects foo, bar and baz like C(baz)->B(bar)->A(foo)->Master(foo,bar,baz) with a window of 2. Change A succeeds but takes too long to replicate, so when the window slides and C's integration jobs start they may pull Master(foo) from a mirror and not actually get A(foo) in the resulting local branch. If A was a fix for integration breakage, then changes which were tested with A in the gate were tested with the fix but changes which started testing immediately after A merged may get tested without the fix and subsequently fail.

Patch-set: 1
Reviewer: Gerrit User 5263 <5263@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Workflow=0