Update patch set 1
Patch Set 1:

All Zuul-triggered jobs should be fetching their commits from Zuul itself, so this shouldn't be a problem.  It's possible some of our post jobs may not be, but they are not triggered by a zuul merge action, they are triggered by gerrit ref-updated events, so this has little bearing on that.  We can probably change them to fetch from zuul as well.

It's possible that repos that are not participating in the current gate stack could pull a different rev because of the timing, but by definition, we don't care about them because they are not participating in the gate stack, and that's already a race condition anyway.  Because they are not in the gate stack, it means they are not in the process of changing and so should be unaffected by such a close timing situation.

Patch-set: 1