Add debugging metrics to RPC
Allow an RPC to get information about the current state of zuul.
This is useful for debugging and monitoring as it is possible to
determine if zuul is failing to hear events from gerrit. Metrics of
the length of the queues can be gathered from the status.json.

This change also exposes the metrics via the webapp as /metrics.json.

Change-Id: Ia649583f2902a7345e0bc59199620a47b5f38851