Update patch set 8
Patch Set 8: Looks good to me (core reviewer)

The test failed because the fake job was waiting, but the WORK_DATA packet that indicated that the job has started was still in-flight, so Zuul didn't know how to cancel the job.

I made a change to the gearman launcher that should accommodate this and rebased this change on that one to hopefully increase its chances of passing tests.

Patch-set: 8
Label: Code-Review=+2
Label: Workflow=0