Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

(1 inline comment)

This is looking pretty good and I'm happy with it. Just a couple of nits:

A test in the test_scheduler.py would be useful. That test framework will set up repositories in /tmp that could be used to ensure a new temp repository is cloned successfully with the refs.

Unimportantly your assertEqual's are in the wrong order. However, so are most of zuuls (should be self.assertEqual(expected, actual, message))

Patch-set: 3
Label: Code-Review=+1