Update patch set 2
Patch Set 2:

So, while I'm here, I wonder what happens if we don't check. Is it possible for gerrit to think the change has merged but it actually hasn't performed the merge? If this is the case will it be gerrit that gets into race conditions with itself as it merges more changes? If gerrit doesn't race against itself (which I'm not sure how that would be possible) then gerrit should catch the problems and it wouldn't be zuul's fault..

Patch-set: 2