Update patch set 3
Patch Set 3: I would prefer that you didn't merge this

(1 inline comment)

This is great, and I can't wait to start using it!

We've found that we need four steps:

#   The zuul ref for the indicated branch
#   The zuul ref for the master branch
#   The tip of the indicated branch
#   The tip of the master branch

Where the indicated branch defaults to ZUUL_BRANCH but can be
overridden by the setup script for cross-branch testing.

So to do something equivalent to the setup_project function in
devstack-gate:

  if self.args.branch:
    override_zuul_ref = re.sub(zuul_branch, self.args.branch, zuul_ref)
  else:
    override_zuul_ref = zuul_ref
  
  if repo.getBranchHead(self.args.branch):
    fallback_zuul_ref = zuul_ref
    fallback_branch = self.args.branch
  else:
    fallback_zuul_ref = re.sub(zuul_branch, 'master', zuul_ref)
    fallback_branch = 'master'
  
  if fetch(override_zuul_ref):
    use this ref
  elif fetch(fallback_zuul_ref):
    use this ref
  else:
    use fallback_branch

Patch-set: 3
Label: Code-Review=-1
Label: Workflow=0