Update patch set 3
Patch Set 3:

Joshua, I will attempt to write a test in the scheduler that uses the cloner. That is a good idea. Will probably start with a very simple test and we can add more later on.

James, your devstack is fairly complicated with fallback branches and such. For this first version I tried to make it useful for a very simple project.  Will look at adding the informations you have identified.

Thank you both for your review on this first version!

Patch-set: 3
Label: Code-Review=0