Update patch set 1
Patch Set 1:

Currently it returns None if status is anything other than MERGED or SUBMITTED. This changes it to instead return None if status is anything other than MERGED. There are already other statuses a change could be in which would be causing this method to return None. I agree that returning False instead of None is probably a good idea, but was trying not to change things too much beyond the intent of the particular bugfix.

Patch-set: 1