Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Okay, I agree that this doesn't change the behaviour and is focused on fixing the bug. I'd be happy for this change to include a correct return from the method or I'm happy with it as it is (and hopefully a subsequent change will tidy it up).

Patch-set: 1
Reviewer: Gerrit User 7069 <7069@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1