Update patch set 1
Patch Set 1: Work In Progress

Hmm, so I did this thinking it would make the code neater and also easier to make future changes to bring zuul's app.js to feature-parity with infra's old-style javascript. However I wonder if the performance difference will be an issue or if this is an improvement at all?

Patch-set: 1