Update patch set 1
Patch Set 1:

The actual change looks fine.

I guess I'm not certain I see the utility of it.  In the past we've focused on using logs to debug problems (we've only had one zuul bug that wasn't completely diagnosable with logs, and in response to that, we added the stack trace signal handler).  Typically we've needed quite a bit of log data to track down zuul bugs; this seems a bit more limited, but maybe I'm just not seeing the use case?  Can you elaborate?

Patch-set: 1
Label: Code-Review=0
Label: Workflow=0