Update patch set 3
Patch Set 3: I would prefer that you didn't merge this

(2 inline comments)

Nearly there :-). Also, as a personal preference, I think it looks neater if we only use the attr on changes that are actually shortened. That way it will remove the unnecessary formatting from changes that don't need to be hovered over to be expanded.

Patch-set: 3
Label: Code-Review=-1