Add subway map to status page
Uses a table method similar to OpenStack's page.

Added a sample JSON file that represents the dependency tree in
a very bad state.  All the images are grey because it's a bit old
and doesn't match the current JSON output in some areas, but it
does exercise quite a bit of the tree code.

Change-Id: I9d9d942f774265d518d624c6a199105c205b8878
Co-Authored-By: James E. Blair <jeblair@openstack.org>