Update patch set 4
Patch Set 4: I would prefer that you didn't merge this

(3 inline comments)

Couple inline things and sorry log blurb here.

I am -1 for a variety of reasons.

First we are adding 10 new config or configuration applying files simply to run two tests (I have not counted the tests themselves here). This is almost impossible to review. There must be a better way to consolidate this stuff and make it readable.

Second less css is problematic for Debian packaging. There are debian packages of zuul. We should stick to css or possibly make lesscpy a possibility for packagers.

TL;DR this is a large complicated change with very little documentation of what is actually going on. I think it should be simplified if possible. And some of the new dependencies are problematic for packagers.

Patch-set: 4
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1
Label: Workflow=0