Fix races in timer trigger tests.
Timer trigger tests are vulnerable to a race where the timer trigger can
enqueue new builds while we are checking the number of builds queued.
This can happen during intermediate build queue checks or in
assertFinalState.

To fix this race swap in a layout that does not use the timer trigger
which will prevent further build queueing once the timer triggered
builds have done their thing.

Change-Id: Ic9b64e34c6a39275b57b1299b47d9b1d0e0e5c79