Fix race in test_idle test.
The test_idle test runs timer triggered jobs in a tight loop. It is
possible for the timer to trigger jobs in the middle of shutting down
the test infrastructure for test_idle. During this shutdown the final
state assertions check that queues are empty but they will not be if the
timer triggered jobs at just the right time.

To fix this race during test cleanup load up a layout without timer
triggered jobs, let any remaining jobs finish running, then end the
test.

Change-Id: Ic9b64e34c6a39275b57b1299b47d9b1d0e0e5c79