Update patch set 15
Patch Set 15: I would prefer that you didn't merge this

This looks good to me, but -1ing it because of the test duplication that you wanted to discuss (specifically what should be a mixin).

So I think we want to test that the mergers are using the correct repo URL rather than the generated gerrit one. Perhaps this isn't even a mixin, but rather a new test that mocks out part of the git merger and checks the remote_url is correct? Then maybe test it with and without remotes etc.

Another test could be to set the remote to somewhere else with a high poll time. Sleep for a few seconds while the merger should be polling then copy the repository/refs into the remote. I think this will test it nicely.

In terms of mixins I think there isn't any harm in using the ones you have thus far in that they exercise fetching certain refs quite well. But I do think they aren't testing the code as well as the above two ideas I've put in. We could still use them and call them something like MergerExcerciseMixin.

Patch-set: 15
Label: Code-Review=-1