Update patch set 16
Patch Set 16:

I don't think we need to worry about testing the getgiturl method without a fetch location set; the part that is currently untested was untested before as well.  I could see adding a simple unit test for it (since it is basically untestable in the current mostly functional test suite).

I do like the idea of exercising the poll code by pushing to the upstream repo after a delay, assuming that's not too complicated of a change.

Patch-set: 16
Label: Code-Review=0
Label: Workflow=0