Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

I'm not entirely sure about putting the merger in the same process; we put the colocated geard option in a separate process for performance reasons (python GIL); perhaps the merger should be as well?  My original thought was that if you wanted to colocate it, you would just start both services on the same host.

OTOH, since processing doesn't have to wait on the merger anymore, maybe having them in the same proc is okay, and maybe we can revisit that for the colocated geard as well?

Unless there's a compelling reason otherwise, I think it would probably be good to use the same pattern for both (eg, both in or out of process, and if in, enabled in the same way (config vs cmdline options)).

Patch-set: 2
Label: Code-Review=-1
Label: Workflow=0