Add option to toggle isJobRegistered check for gearman
Currently, zuul checks to see if a job is registerd with gearman.
This is useful when a job is always created in Jenkins however it
is possible for jobs to be valid, but not registered with Jenkins.

Now we have the ability to toggle this check, and assume the job is
actually registered.

Closes-Bug: #1302756

Change-Id: Iab6a1d06f1ed9fd0ee44f64c622f9e24832a24f4
Signed-off-by: Paul Belanger <paul.belanger@polybeacon.com>