Pass change and pipeline to the status_url
Our JavaScript evangelist Timo Tijhof would like to highlight a given
change on the Zuul status page.  To do so, the Zuul status URL reported
in Gerrit would be suffixed with an anchor followed by the change and
patchset number.

This patch uses format() on the status_url and pass it the change and
pipeline objects. Thus with:

 status_url=http://zuul.example.com/#{change.number},{change.patchset}

We would end up reporting in Gerrit:

  Starting gate jobs.
  https://zuul.example.com/#12345,1

Some JavaScript on the status page would do the rest and highlight the
corresponding patch.

The test is based on TestScheduler.test_check_smtp_pool().

Downstream bug for reference:
  https://bugzilla.wikimedia.org/63744

Change-Id: Ia34de996d2c58620adf1e6e39b97266e867e2092