Update patch set 4
Patch Set 4: I would prefer that you didn't merge this

I think part of the trouble I'm having with this is that each one of those metrics isn't really a metric -- they are internal implementation details which might be useful for debugging, but not actually useful for identifying a failure externally.  And I don't think they should be part of the contract between Zuul and the operator.  For example, we should never break someone's monitoring system because we get rid of the layout lock.

I fully agree that systems should be monitored, which is why Zuul emits copious statsd events.  In fact, I'm certain that you can get the time of the last Gerrit event seen by Zuul from graphite (which can produce data not only in graph form, but also JSON for easy programmatic parsing).

If you really don't want to use graphite (or statsd) and would rather add this approach, then I think it should focus on aspects that the operator is more likely to care about (such as timestamp of last event, count of events, etc).  Also, I'm not really keen on the Heisenberg effect some of the items in the current example have (eg, keeping event objects around long after they would normally be destroyed just in case they are used in a metric report later).

Patch-set: 4
Label: Code-Review=-1