Gerrit reporter can message without parameters
The Zuul Gerrit reporter would returns early whenever it is missing a
parameter, the workaround was to pass force-message: True to report the
message.

In Gerrit 2.8 the --force-message option of 'gerrit review' has been
removed, since you can now comment on a closed change (votes are discard
though in such a case).

To have a pipeline to never report back, simply do not specify a
Gerrit reporter for it.

Reference:
    https://gerrit-review.googlesource.com/#/c/45441/
    commit c2713185

Change-Id: Ia9ca74a2a6eca6dbf39e468753fa0bcb6fcc0eba