Make zuul javascript app more modular
Refactor the zuul status app to be more modular. This allows the
controlling page (eg index.html) to load in only the components
it requires. This will make it easier to format just one patchset
or change where the layout controls are etc.

This is to work towards overlaying zuul results in gerrit.

Change-Id: Idfba9f08c25850c340cc962fa0f5f9652306b877