Skip closed changes
If a change is closed, zuul should not bother to run any jobs for
it. Gerrit in particular will still allow commenting on closed
changes, but zuul will be unable to vote on them and so there is
little value in running jobs for them in the event a comment-added
event is received which might otherwise trigger some action.

Change-Id: I6b38596cf7158290cb348aba12f3eacf49b87ad0