Implement Depends-On: IdempotentID
This is a first pass implementation of the Depends-On: tag in zuul.

The semantics are that Depends-On: acts exactly like a git
dependency. We look at a change for it's Depends-On tags and if so
treat it like a zuul dependency. We also forward search gerrit to
pull out the neededby based on the Idempotent id.

Change-Id: Ib207ef86583bf635ed6768de6cead6dcd893b012