Update patch set 4
Patch Set 4:

Josh, this change is independent on whether openstack supports drafts (or not).  The only purpose of this change is to make zuul aware of draft events (via the gerrit event stream).  It's in the works to disable drafts on review.o.o but other organizations using zuul & gerrit may want to support drafts so they may want this change.

Patch-set: 4