Make swiftclient an optional dependency
I can not use the switclient on my installation and need a way to
prevent the swiftclient import entirely.

The only usage of the swiftclient module is to establish a connection
which only happen if there is a [swift] section in the configuration.

Moving the import from the top of the file next to its only usage would
make sure my installs never has to get swiftclient installed.

Change-Id: Ie2ef84f86f9f57796a85758147ba87d8b22c3e55