Update patch set 1
Patch Set 1: Code-Review-1

This is great thank you. We should also document that SIGUSR2 toggles the state of yappi.So on the first SIGUSR2 we enable yappi profiling. On the second SIGUSR2 we dump the information collected, reset all yappi state and stop profiling. This decision was made to minimize the impact of yappi on a running system.

Patch-set: 1
Label: Code-Review=-1