Update patch set 20
Patch Set 20:

Made it to clone the repositories in the same order the repositories are passed as argument.  The clone mapper was maintaining an unsorted dict which was later sorted and cause git clone to complain when the destination already exist but is not a git repository.

https://review.openstack.org/#/c/70373/19..20/doc/source/cloner.rst,unified
https://review.openstack.org/#/c/70373/19..20/zuul/lib/clonemapper.py

Patch-set: 20
Label: Verified=0