Add support for negative requirements (take 2)
This change was originally reverted due to confusion over the
terminology and expectations. Nothing has changed in this patch, it
is just re-applying the reverted version. This is open to comments
for how it should be functioning (and what terminology should be
used). Therefore, it is WIP.

This change adds support for false matching of requirements. To make
this useful you can now require all approvals match a requirement or
only one (ie any).

Therefore depreciate require-approvals, replacing with
require-approvals-any and a new require-approvals-all.

Change-Id: I3369920530e0b7439208b8fd43a9e75994860666