WIP Add support for negative requirements
This change adds support for false matching of requirements. To make
this useful you can now require all approvals match a requirement or
only one (ie any).

Therefore depreciate require-approvals, replacing with
require-approvals-any and a new require-approvals-all.

This is WIP as it needs more tests and is just a first pass.

(Depends on German beer from Jogo)

Change-Id: I458e677315ccb90d64cd0c0e734951141324a9c3