Update patch set 21
Patch Set 21: Code-Review-1

(3 comments)

Just some minor things inline, otherwise, looks really good, thanks!

I think we've added one more kind of fallback to devstack-gate since this was written, but I'd be happy to take the task of adding support for that in another patchset if you don't want to.  Also, there are some good unit tests there that I think we can expand this with, but again, that can be another patch.

Patch-set: 21
Label: Code-Review=-1