Update patch set 23
Patch Set 23:

I looked at the per project branch override support in a child change. The code is ugly and does not quite work  https://review.openstack.org/#/c/104929/ 

Need to figure out something simpler and more robust. I am unlikely to be able to polish it up this week or next week though  :-/

Even with the follow up change. The Zuul cloner will not be able to match devstack-gate expectations. But that is already a good base for the simplest cases :-D

Patch-set: 23